+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Version History
 
-## Unreleased
+## Development
+
+### Removed
+
+* Removed dependency on `casacore` and `ska-sdp-metadata-generator` module.
+
+### Changed
+
+* Updated the `requests.post` call in the client `register_data_item` function to send metadata as JSON in the request body.
 
 ## 0.2.0